requirements.txt
django==4.2.7
djangorestframework==3.14.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
coreapi==2.3.3
django-cors-headers==4.3.1
boto3==1.34.69
djongo==1.3.6
pymongo==3.12.3
pandas==2.2.3
openpyxl==3.1.5